httpd: new reverse proxy handling (#94)
Webtop `5.11.0` introduce a new auth log where are recorded all the access
attempt and the associated information, like sessions, location (if possible)
and source IP address. To determine the source IP address, WebTop inspects the
header `X-Forwarded-For` if it is present, otherwise falls back on the request
IP address.

Due to a bug in the Centos provided Apache HTTP version `2.4.6`, if `ProxyPass`
is used to forward the request to WebTop, after some time it stops adding the
header `X-Forwarded-For` into the requests. As a consequence of this bug, Webtop
starts to report `127.0.0.1` as the IP source of the authentication attempts.

A workaround for this bug is:

* Rewrite the proxy directive using `RewriteRule` with the `P` flag instead
  of `ProxyPass`
* Move part of the proxy configuration in a `.htaccess` file

Co-authored-by: Giacomo Sanchietti <giacomo.sanchietti@nethesis.it>
Issue: NethServer/dev#6463